refine admin dashboard flow
