+++ apps/TOM-web/src/app/admin/useAdminDashboard.ts
@@ -1,0 +1,141 @@
+'use client';
+
+import { useState } from 'react';
+
+import {
+  createPendingRequest,
+  formatThresholdLabel,
+  initialActiveClubs,
+  initialForm,
+  initialRequests,
+  initialSpamQueue,
+  requestToActiveClub,
+  type ClubForm,
+  type ClubRequest,
+  thresholdGoal,
+} from './admin-data';
+
+export function useAdminDashboard() {
+  const [form, setForm] = useState(initialForm);
+  const [requests, setRequests] = useState(initialRequests);
+  const [activeClubs, setActiveClubs] = useState(initialActiveClubs);
+  const [spamQueue, setSpamQueue] = useState(initialSpamQueue);
+  const [banner, setBanner] = useState(
+    'Шинэ клубийн хүсэлтүүдийг шалгаж, боломжтойг нь идэвхжүүлж, spam-ийг хурдан цэвэрлэ.'
+  );
+
+  const updateField = (field: keyof ClubForm, value: string) => {
+    setForm((current) => ({
+      ...current,
+      [field]: value,
+    }));
+  };
+
+  const resetForm = () => {
+    setForm(initialForm);
+  };
+
+  const handleCreate = () => {
+    const newRequest: ClubRequest = createPendingRequest(form);
+
+    setRequests((current) => [newRequest, ...current]);
+    setBanner(`${newRequest.clubName} шалгах дараалалд нэмэгдлээ.`);
+    setForm(initialForm);
+  };
+
+  const approveRequest = (requestId: string) => {
+    const request = requests.find((item) => item.id === requestId);
+
+    if (!request) return;
+
+    setRequests((current) =>
+      current.map((item) =>
+        item.id === requestId
+          ? { ...item, requestStatus: 'approved', clubStatus: 'active' }
+          : item
+      )
+    );
+
+    setActiveClubs((current) => {
+      const nextClub = requestToActiveClub(request);
+      return [nextClub, ...current.filter((item) => item.id !== requestId)];
+    });
+    setBanner(`${request.clubName} батлагдаж active төлөвт шилжлээ.`);
+  };
+
+  const rejectRequest = (requestId: string) => {
+    const request = requests.find((item) => item.id === requestId);
+
+    if (!request) return;
+
+    setRequests((current) =>
+      current.map((item) =>
+        item.id === requestId
+          ? { ...item, requestStatus: 'rejected', clubStatus: 'paused' }
+          : item
+      )
+    );
+    setActiveClubs((current) =>
+      current.filter((item) => item.id !== requestId)
+    );
+    setBanner(`${request.clubName} татгалзагдлаа.`);
+  };
+
+  const toggleClubStatus = (clubId: string) => {
+    setActiveClubs((current) =>
+      current.map((club) => {
+        if (club.id !== clubId) {
+          return club;
+        }
+
+        const nextStatus = club.clubStatus === 'active' ? 'paused' : 'active';
+        return {
+          ...club,
+          clubStatus: nextStatus,
+          requestStatus: nextStatus === 'active' ? 'approved' : club.requestStatus,
+        };
+      })
+    );
+    setBanner('Клубийн төлөв шинэчлэгдлээ.');
+  };
+
+  const removeSpamClub = (clubId: string) => {
+    const spamClub = spamQueue.find((item) => item.id === clubId);
+
+    if (!spamClub) return;
+
+    setSpamQueue((current) => current.filter((item) => item.id !== clubId));
+    setBanner(`${spamClub.clubName} spam гэж устгагдлаа.`);
+  };
+
+  const pendingRequests = requests.filter(
+    (request) => request.requestStatus === 'pending'
+  );
+  const activeCount = activeClubs.filter((club) => club.clubStatus === 'active')
+    .length;
+  const thresholdReachedCount = new Set(
+    [...requests, ...activeClubs]
+      .filter((club) => club.interestCount >= thresholdGoal)
+      .map((club) => club.id)
+  ).size;
+
+  return {
+    activeClubs,
+    activeCount,
+    approveRequest,
+    banner,
+    form,
+    handleCreate,
+    pendingRequests,
+    rejectRequest,
+    removeSpamClub,
+    requests,
+    spamQueue,
+    thresholdReachedCount,
+    thresholdGoal,
+    resetForm,
+    toggleClubStatus,
+    updateField,
+    formatThresholdLabel,
+  };
+}
